refactor: simplify vendor dashboard layout by moving components to `Layout` and removing unused imports, optimize avatar hover styling
